+++ src/reid_score/__init__.py
@@ -1,6 +1,6 @@
 """reid_score public API."""
 
-from .scorer import ReidScorer
+from .scorer import BatchScoringError, ReidScorer
 from .types import (
     CompareResult,
     InferredAttribute,
@@ -11,6 +11,7 @@ from .types import (
 
 __all__ = [
     "ReidScorer",
+    "BatchScoringError",
     "ReidConfig",
     "InferredAttribute",
     "ScoreResult",